refactor(chat): extract bottom-sheet chat message rendering into components
